refactor: global/inmemory 디렉토리를 infra/inmemory로 이동
